Align naming and directory structure with other k8s operators
