+++ pkg/operator/targetconfigcontroller/targetconfigcontroller.go
@@ -1,24 +1,33 @@
-package operator
+package targetconfigcontroller
 
 import (
 	"context"
 	"fmt"
 	"sort"
 	"strings"
+	"time"
 
-	"github.com/openshift/cluster-kube-scheduler-operator/pkg/operator/operatorclient"
-	"github.com/openshift/cluster-kube-scheduler-operator/pkg/operator/v410_00_assets"
-	"github.com/openshift/cluster-kube-scheduler-operator/pkg/version"
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
+	"k8s.io/apimachinery/pkg/util/sets"
+	"k8s.io/apimachinery/pkg/util/wait"
+	"k8s.io/client-go/informers"
+	"k8s.io/client-go/kubernetes"
 	corev1client "k8s.io/client-go/kubernetes/typed/core/v1"
 	corev1listers "k8s.io/client-go/listers/core/v1"
+	"k8s.io/client-go/tools/cache"
+	"k8s.io/client-go/util/workqueue"
 	"k8s.io/klog"
 
 	"github.com/openshift/api/config/v1"
 	operatorv1 "github.com/openshift/api/operator/v1"
+	configinformers "github.com/openshift/client-go/config/informers/externalversions"
 	configlistersv1 "github.com/openshift/client-go/config/listers/config/v1"
+	"github.com/openshift/cluster-kube-scheduler-operator/pkg/operator/operatorclient"
+	"github.com/openshift/cluster-kube-scheduler-operator/pkg/operator/v410_00_assets"
+	"github.com/openshift/cluster-kube-scheduler-operator/pkg/version"
 	"github.com/openshift/library-go/pkg/operator/events"
 	"github.com/openshift/library-go/pkg/operator/resource/resourceapply"
 	"github.com/openshift/library-go/pkg/operator/resource/resourcemerge"
@@ -27,11 +36,193 @@ import (
 	"github.com/openshift/library-go/pkg/operator/v1helpers"
 )
 
-const TargetPolicyConfigMapName = "policy-configmap"
+const (
+	workQueueKey              = "key"
+	TargetPolicyConfigMapName = "policy-configmap"
+)
 
-// syncKubeScheduler_v311_00_to_latest takes care of synchronizing (not upgrading) the thing we're managing.
+type TargetConfigController struct {
+	ctx                   context.Context
+	targetImagePullSpec   string
+	operatorImagePullSpec string
+	operatorClient        v1helpers.StaticPodOperatorClient
+	kubeClient            kubernetes.Interface
+	eventRecorder         events.Recorder
+	configMapLister       corev1listers.ConfigMapLister
+	featureGateLister     configlistersv1.FeatureGateLister
+	featureGateCacheSync  cache.InformerSynced
+	// queue only ever has one item, but it has nice error handling backoff/retry semantics
+	queue workqueue.RateLimitingInterface
+}
+
+func NewTargetConfigController(
+	ctx context.Context,
+	targetImagePullSpec, operatorImagePullSpec string,
+	operatorConfigClient v1helpers.OperatorClient,
+	namespacedKubeInformers informers.SharedInformerFactory,
+	kubeInformersForNamespaces v1helpers.KubeInformersForNamespaces,
+	configInformer configinformers.SharedInformerFactory,
+	operatorClient v1helpers.StaticPodOperatorClient,
+	kubeClient kubernetes.Interface,
+	eventRecorder events.Recorder,
+) *TargetConfigController {
+	c := &TargetConfigController{
+		ctx:                   ctx,
+		targetImagePullSpec:   targetImagePullSpec,
+		operatorImagePullSpec: operatorImagePullSpec,
+		kubeClient:            kubeClient,
+		configMapLister:       kubeInformersForNamespaces.ConfigMapLister(),
+		operatorClient:        operatorClient,
+		eventRecorder:         eventRecorder,
+
+		featureGateLister:    configInformer.Config().V1().FeatureGates().Lister(),
+		featureGateCacheSync: configInformer.Config().V1().FeatureGates().Informer().HasSynced,
+		queue:                workqueue.NewNamedRateLimitingQueue(workqueue.DefaultControllerRateLimiter(), "TargetConfigController"),
+	}
+
+	operatorConfigClient.Informer().AddEventHandler(c.eventHandler())
+	namespacedKubeInformers.Core().V1().ConfigMaps().Informer().AddEventHandler(c.eventHandler())
+	namespacedKubeInformers.Core().V1().Secrets().Informer().AddEventHandler(c.eventHandler())
+	namespacedKubeInformers.Core().V1().ServiceAccounts().Informer().AddEventHandler(c.eventHandler())
+
+	configInformer.Config().V1().FeatureGates().Informer().AddEventHandler(c.eventHandler())
+	// we react to some config changes
+	kubeInformersForNamespaces.InformersFor(operatorclient.GlobalUserSpecifiedConfigNamespace).Core().V1().ConfigMaps().Informer().AddEventHandler(c.eventHandler())
+
+	// we only watch some namespaces
+	namespacedKubeInformers.Core().V1().Namespaces().Informer().AddEventHandler(c.namespaceEventHandler())
+
+	return c
+}
+
+func (c TargetConfigController) sync() error {
+	operatorSpec, _, _, err := c.operatorClient.GetStaticPodOperatorState()
+	if err != nil {
+		return err
+	}
+
+	switch operatorSpec.ManagementState {
+	case operatorv1.Managed:
+	case operatorv1.Unmanaged:
+		return nil
+
+	case operatorv1.Removed:
+		// TODO probably just fail
+		return nil
+	default:
+		c.eventRecorder.Warningf("ManagementStateUnknown", "Unrecognized operator management state %q", operatorSpec.ManagementState)
+		return nil
+	}
+	requeue, err := createTargetConfigController_v311_00_to_latest(c, c.eventRecorder, operatorSpec)
+	if err != nil {
+		return err
+	}
+	if requeue {
+		return fmt.Errorf("synthetic requeue request")
+	}
+
+	return nil
+}
+
+// Run starts the kube-scheduler and blocks until stopCh is closed.
+func (c *TargetConfigController) Run(workers int, stopCh <-chan struct{}) {
+	defer utilruntime.HandleCrash()
+	defer c.queue.ShutDown()
+
+	klog.Infof("Starting TargetConfigController")
+	defer klog.Infof("Shutting down TargetConfigController")
+
+	if !cache.WaitForCacheSync(stopCh, c.featureGateCacheSync) {
+		utilruntime.HandleError(fmt.Errorf("caches did not sync"))
+		return
+	}
+
+	// doesn't matter what workers say, only start one.
+	go wait.Until(c.runWorker, time.Second, stopCh)
+
+	<-stopCh
+}
+
+func (c *TargetConfigController) runWorker() {
+	for c.processNextWorkItem() {
+	}
+}
+
+func (c *TargetConfigController) processNextWorkItem() bool {
+	dsKey, quit := c.queue.Get()
+	if quit {
+		return false
+	}
+	defer c.queue.Done(dsKey)
+
+	err := c.sync()
+	if err == nil {
+		c.queue.Forget(dsKey)
+		return true
+	}
+
+	utilruntime.HandleError(fmt.Errorf("%v failed with : %v", dsKey, err))
+	c.queue.AddRateLimited(dsKey)
+
+	return true
+}
+
+// eventHandler queues the operator to check spec and status
+func (c *TargetConfigController) eventHandler() cache.ResourceEventHandler {
+	return cache.ResourceEventHandlerFuncs{
+		AddFunc:    func(obj interface{}) { c.queue.Add(workQueueKey) },
+		UpdateFunc: func(old, new interface{}) { c.queue.Add(workQueueKey) },
+		DeleteFunc: func(obj interface{}) { c.queue.Add(workQueueKey) },
+	}
+}
+
+// this set of namespaces will include things like logging and metrics which are used to drive
+var interestingNamespaces = sets.NewString(operatorclient.TargetNamespace)
+
+func (c *TargetConfigController) namespaceEventHandler() cache.ResourceEventHandler {
+	return cache.ResourceEventHandlerFuncs{
+		AddFunc: func(obj interface{}) {
+			ns, ok := obj.(*corev1.Namespace)
+			if !ok {
+				c.queue.Add(workQueueKey)
+			}
+			if ns.Name == operatorclient.TargetNamespace {
+				c.queue.Add(workQueueKey)
+			}
+		},
+		UpdateFunc: func(old, new interface{}) {
+			ns, ok := old.(*corev1.Namespace)
+			if !ok {
+				c.queue.Add(workQueueKey)
+			}
+			if ns.Name == operatorclient.TargetNamespace {
+				c.queue.Add(workQueueKey)
+			}
+		},
+		DeleteFunc: func(obj interface{}) {
+			ns, ok := obj.(*corev1.Namespace)
+			if !ok {
+				tombstone, ok := obj.(cache.DeletedFinalStateUnknown)
+				if !ok {
+					utilruntime.HandleError(fmt.Errorf("couldn't get object from tombstone %#v", obj))
+					return
+				}
+				ns, ok = tombstone.Obj.(*corev1.Namespace)
+				if !ok {
+					utilruntime.HandleError(fmt.Errorf("tombstone contained object that is not a Namespace %#v", obj))
+					return
+				}
+			}
+			if ns.Name == operatorclient.TargetNamespace {
+				c.queue.Add(workQueueKey)
+			}
+		},
+	}
+}
+
+// createTargetConfigController_v311_00_to_latest takes care of synchronizing (not upgrading) the thing we're managing.
 // most of the time the sync method will be good for a large span of minor versions
-func createTargetConfigReconciler_v311_00_to_latest(c TargetConfigReconciler, recorder events.Recorder, operatorSpec *operatorv1.StaticPodOperatorSpec) (bool, error) {
+func createTargetConfigController_v311_00_to_latest(c TargetConfigController, recorder events.Recorder, operatorSpec *operatorv1.StaticPodOperatorSpec) (bool, error) {
 	errors := []error{}
 
 	_, _, err := manageKubeSchedulerConfigMap_v311_00_to_latest(c.configMapLister, c.kubeClient.CoreV1(), recorder, operatorSpec)
