+++ pkg/operator/targetconfigcontroller/targetconfigcontroller_test.go
@@ -1,4 +1,4 @@
-package operator
+package targetconfigcontroller
 
 import (
 	"reflect"
